transformed canMaintainSymlinkTimestamps into a function and relocated unix specific tests
