+++ target_disk.go
@@ -130,7 +130,7 @@ func (d *TargetDisk) Chown(name string, uid, gid int) error {
 
 // Lchtimes changes the access and modification times of the named file.
 func (d *TargetDisk) Lchtimes(name string, atime, mtime time.Time) error {
-	if canMaintainSymlinkTimestamps() {
+	if canMaintainSymlinkTimestamps {
 		return lchtimes(name, atime, mtime)
 	}
 	return nil
