+++ target_disk_unix.go
@@ -35,6 +35,4 @@ func unixTimeval(t time.Time) unix.Timeval {
 // packages (Chmod and Chtimes) are not capable of changing symlink info
 // because those methods follow the symlinks. However, a platform-dependent option
 // is provided for unix (see Lchtimes)
-func canMaintainSymlinkTimestamps() bool {
-	return true
-}
+const canMaintainSymlinkTimestamps = true
